Changed ID data types

+++ IDB/Load/BCP/IDB.Content/File.cs
@@ -5,16 +5,16 @@ namespace IDB.Load.BCP.IDB.Content
     class File
     {
         private string _name;
-        private int _id;
+        private long _id;
         private string _localPath;
         private DateTime _date;
         private string _iterationId;
         internal string Name { set { _name = value; } get { return _name; } }
         internal string LocalPath { set { _localPath = value; } get { return _localPath; } }
-        internal int Id { set { _id = value; } get { return _id; } }
+        internal long Id { set { _id = value; } get { return _id; } }
         internal string IterationId { set { _iterationId = value; } get { return _iterationId; } }
         internal DateTime Date { set { _date = value; } get { return _date; } }
-      internal  File(int fileId,string fileName, string fileLocalPath, DateTime createdDate,string fileIterationId) 
+      internal  File(long fileId,string fileName, string fileLocalPath, DateTime createdDate,string fileIterationId) 
         {
             _iterationId = fileIterationId;
             _id = fileId;
